requirements.txt
pandas==1.5.2
PyQt5==5.15.7
requests==2.28.1
tabulate==0.9.0
web3==5.31.1
python-binance==1.0.16
